requirements.txt
gradio>=6.10.0
groq>=1.1.0
python-dotenv>=1.2.0
